refactor(mocks): extract chat module from mockApi (#991 M3)
Move the chat helpers (CHAT_MAX_TURNS, chatDTO, truncateChatTitle) and the ten
chat methods (listChats, createChat, sendChatMessage, endChat, continueChat,
confirmProposal, dismissProposal, startRunFromChat, cancelRunFromChat,
steerRunFromChat) verbatim into web/src/mocks/mockApi/chat.ts as chatApi.

D7 (the one intended non-motion edit in all of M3): inside startRunFromChat the
call `return mockApi.createRun(repoId, card.iid);` becomes
`return runsApi.createRun(repoId, card.iid);`, and chat.ts imports
`{ runsApi } from "./runs"`. The composed mockApi object does not exist from
inside a partial; fact 1 measured this is the only intra-object call and fact 7
measured no test spies on createRun through the object, so behaviour is
preserved. This is exactly one line.

noUnusedLocals then dropped from index.ts the whole ../engine import
(handleInput/scheduleChatReply), the appendMessage/getProposal/getRun/putProposal
bindings from ../store, and the Chat/CreatedIssue/Run/RunMessage types.
